src/requirements.txt
flask==2.2.2
flasgger==0.9.5
flask_marshmallow==0.14.0
apispec==6.0.2
gevent==22.10.2
Flask-Permissions==0.2.3
Flask-Login==0.6.2
psycopg2==2.9.5
flask_sqlalchemy==3.0.2
redis==4.3.5
sqlalchemy==1.4.44
pydantic[dotenv]==1.9.0
Flask-JWT==0.3.2
flask-redis==0.4.0
pbkdf2==1.3
aioredis==2.0.1
marshmallow_enum
gunicorn[gevent]==20.1.0
SQLAlchemy-Paginator==0.2
Flask-Migrate==4.0.0
psycogreen==1.0.2